feat(ci): add Rust modules build CI workflow and update .gitignore
- Add GitHub Actions workflow for building Rust modules (xray-signal-handler, pepper-shaper, perf-net, quiche-client)
- Build for all Android targets: aarch64, armv7, x86_64, i686
- Add Rust build artifacts to .gitignore (target/, *.so, *.a files)
- Configure caching for Rust dependencies and build artifacts

+++ app/src/main/jni/pepper-shaper/src/pacing.rs
@@ -102,7 +102,7 @@ pub fn update_after_send(
     // Update next send time based on rate
     if params.target_rate_bps > 0 {
         // Calculate interval: packet_size * 8 bits / rate_bps = seconds
-        let interval_ns = (packet_size * 8 * 1_000_000_000) / params.target_rate_bps;
+        let interval_ns = ((packet_size as u64) * 8 * 1_000_000_000) / params.target_rate_bps;
         state.next_send_time_ns = current_time_ns + interval_ns.max(params.min_pacing_interval_ns);
     } else {
         // No rate limit, just respect min interval
